+++ spec/dummy/client/entries/app-kitten.js
@@ -44,6 +44,7 @@ import { TextInput } from 'kitten/components/form/text-input'
 import { LocationInput } from 'kitten/components/form/location-input'
 import { TextInputWithLimit } from 'kitten/components/form/text-input-with-limit'
 import { TextInputWithUnit } from 'kitten/components/form/text-input-with-unit'
+import { TextInputWithButton } from 'kitten/components/form/text-input-with-button'
 
 // Grid
 import { Container } from 'kitten/components/grid/container'
@@ -76,6 +77,7 @@ import { InfoLines } from 'kitten/components/lines/info-lines'
 
 // Links
 import ExternalRichLink from 'kitten/components/links/external-rich-link'
+import { LinkList } from 'kitten/components/links/link-list'
 
 // Lists
 import { BulletList } from 'kitten/components/lists/bullet-list'
@@ -246,6 +248,7 @@ ReactOnRails.register({
   TextInput,
   TextInputWithLimit,
   TextInputWithUnit,
+  TextInputWithButton,
 
   // Grid
   Container,
@@ -278,6 +281,7 @@ ReactOnRails.register({
 
   // Links
   ExternalRichLink,
+  LinkList,
 
   // Lists
   BulletList,