+++ spec/dummy/client/entries/app-kitten.js
@@ -23,6 +23,7 @@ import { TagButton } from 'kitten/components/buttons/tag-button'
 // Cards
 import { ProjectCard } from 'kitten/components/cards/project-card'
 import { ProjectCreatorCard } from 'kitten/components/cards/project-creator-card'
+import { ProjectSimilarCard } from 'kitten/components/cards/project-similar-card'
 
 // Dev
 import { DevGrid } from 'kitten/components/dev/dev-grid'
@@ -189,6 +190,8 @@ import { KarlIconBadgeWithSvg } from 'kitten/karl/notifications/icon-badge'
 import { KarlLegoGrid } from 'kitten/karl/grid/lego-grid'
 import { KarlTagList } from 'kitten/karl/lists/tag-list'
 import { KarlCard } from 'kitten/karl/cards/card'
+import { KarlProjectSimilarCard,
+         KarlProjectSimilarCardInfos } from 'kitten/karl/cards/project-similar-card'
 import { KarlMenuHeader } from 'kitten/karl/headers/karl-header'
 import { KarlGithubRibbon } from 'kitten/karl/headers/karl-github-ribbon'
 import { KarlMenu,
@@ -217,6 +220,7 @@ ReactOnRails.register({
   // Cards
   ProjectCard,
   ProjectCreatorCard,
+  ProjectSimilarCard,
 
   // Dev
   Checkbox,
@@ -379,6 +383,8 @@ ReactOnRails.register({
   KarlLegoGrid,
   KarlTagList,
   KarlCard,
+  KarlProjectSimilarCard,
+  KarlProjectSimilarCardInfos,
   KarlMenuHeader,
   KarlMenu,
   KarlMenuContent,