+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add new `ProjectSimilarCard` component.
+- Feature: Add new `RefreshIcon` component.
 
 ## [13.3.1] - 2017-07-06