+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add animation `ButtonIcon` on `ProjectCreatorCard`.
+
 ## [13.3.1] - 2017-07-06
 
 Fix: