+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `LinkList` component.
+
 ## [13.3.1] - 2017-07-06
 
 Fix: