+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add new `LinkList` component.
+- Fix: Add `withoutHover` and `verticalArrow` options on `ButtonIcon` component.
+- Fix: Add `ButtonIcon` animation on `ProjectCreatorCard`.
 - Feature: Add `k-u-align-center-block` utility to center blocks.
 
 ## [13.3.1] - 2017-07-06