+++ CHANGELOG.md
@@ -6,7 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add `Loader` and `LoaderWithParagraph` components.
 - Feature: Add `icon` and `iconWithMinWidth` props to `Button` component.
+- Feature: Add new `LinkList` component.
+- Feature: Add new `TextInputWithButton` component.
 - Feature: Add `k-u-align-center-block` utility to center blocks.
+- Fix: Add `withoutHover` and `verticalArrow` options on `ButtonIcon` component.
+- Fix: Add `ButtonIcon` animation on `ProjectCreatorCard`.
 
 ## [13.3.1] - 2017-07-06