+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add social button icon components (`FacebookButtonIcon`, …).
+- Feature: Add new `ProjectSimilarCard` component.
+- Feature: Add new `RefreshIcon` component.
 
 ## [13.4.0] - 2017-07-20