+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `Modal` component.
+- Feature: Add `Loader` and `LoaderWithParagraph` components.
+- Feature: Add `icon` and `iconWithMinWidth` props to `Button` component.
+- Feature: Add social button icon components (`FacebookButtonIcon`, …).
+- Feature: Add border utilities (`k-u-border-left-line1`, …).
+- Feature: Add new `ProjectSimilarCard` component.
+- Feature: Add new `RefreshIcon` component.
 
 ## [13.4.0] - 2017-07-20